doxygen docs added
LOTS and LOTS of files added as part of the doxygen doc tool. IDK why I
tried this, but now I will push no matter what, because it took me
effort and the documentation does work, whether or not anybody has
written it properly. Also, good luck pulling this commit.